+++ lib/TokenManager.ts
@@ -26,7 +26,8 @@ import {
   isRefreshToken,
   StorageOptions,
   StorageType,
-  OktaAuth
+  OktaAuth,
+  RefreshToken
 } from './types';
 import { ID_TOKEN_STORAGE_KEY, ACCESS_TOKEN_STORAGE_KEY, REFRESH_TOKEN_STORAGE_KEY } from './constants';
 
@@ -162,6 +163,15 @@ function validateToken(token: Token) {
   }
 }
 
+function _add(sdk, tokenMgmtRef, storage, key, token: Token) {
+  var tokenStorage = storage.getStorage();
+  validateToken(token);
+  tokenStorage[key] = token;
+  storage.setStorage(tokenStorage);
+  emitAdded(tokenMgmtRef, key, token);
+  setExpireEventTimeout(sdk, tokenMgmtRef, key, token);
+}
+
 function add(sdk, tokenMgmtRef, storage, key, token: Token) {
   if (sdk.features.isLocalhost()) {
     warn(
@@ -171,12 +181,7 @@ function add(sdk, tokenMgmtRef, storage, key, token: Token) {
     );
   }
 
-  var tokenStorage = storage.getStorage();
-  validateToken(token);
-  tokenStorage[key] = token;
-  storage.setStorage(tokenStorage);
-  emitAdded(tokenMgmtRef, key, token);
-  setExpireEventTimeout(sdk, tokenMgmtRef, key, token);
+  _add(sdk, tokenMgmtRef, storage, key, token);
 }
 
 function get(storage, key) {
@@ -300,6 +305,24 @@ function remove(tokenMgmtRef, storage, key) {
   emitRemoved(tokenMgmtRef, key, removedToken);
 }
 
+function _renewToken(sdk, token) {
+  return this.getTokens()
+    .then(tokens => tokens.refreshToken as RefreshToken)
+    .then(refreshTokenObject => {
+      if (refreshTokenObject) {
+        return sdk.token.renewTokensWithRefresh({
+          scopes: token.scopes,
+        }, refreshTokenObject)
+        .then(function(freshTokens) {
+          // Multiple tokens may have come back. Return only the token which was requested.
+          return isIDToken(token) ? freshTokens.idToken : freshTokens.accessToken;
+        });
+      } else {
+        return sdk.token.renew(token);
+      }
+    });
+}
+
 function renew(sdk, tokenMgmtRef, storage, key) {
   // Multiple callers may receive the same promise. They will all resolve or reject from the same request.
   var existingPromise = tokenMgmtRef.renewPromise[key];
@@ -316,59 +339,36 @@ function renew(sdk, tokenMgmtRef, storage, key) {
     return Promise.reject(e);
   }
 
-  // Remove existing autoRenew timeouts
-  clearExpireEventTimeoutAll(tokenMgmtRef);
+  // Remove existing autoRenew timeout
+  clearExpireEventTimeout(tokenMgmtRef, key);
 
   // A refresh token means a replace instead of renewal
-
   // Store the renew promise state, to avoid renewing again
-  // Renew/refresh all tokens in one process
-  tokenMgmtRef.renewPromise[key] = sdk.token.renewTokens({
-    scopes: token.scopes,
-  })
-    .then(function(freshTokens) {
-      // store and emit events for freshTokens
+  tokenMgmtRef.renewPromise[key] = _renewToken.call(this, sdk, token)
+    .then(function(freshToken) {
+      // store and emit events for freshToken
       const oldTokenStorage = storage.getStorage();
-      setTokens(
-        sdk, 
-        tokenMgmtRef, 
-        storage, 
-        freshTokens, 
-        (accessTokenKey, accessToken) =>
-          emitRenewed(tokenMgmtRef, accessTokenKey, accessToken, oldTokenStorage[accessTokenKey]),
-        (idTokenKey, idToken) =>
-          emitRenewed(tokenMgmtRef, idTokenKey, idToken, oldTokenStorage[idTokenKey]),
-        // not emitting refresh token as an internal detail, not a usable token
-      );
-
-      // return freshToken by key
-      const freshToken = get(storage, key);
+      remove(tokenMgmtRef, storage, key);
+      _add(sdk, tokenMgmtRef, storage, key, freshToken);
+      emitRenewed(tokenMgmtRef, key, freshToken, oldTokenStorage[key]);
       return freshToken;
     })
     .catch(function(err) {
       if (err.name === 'OAuthError' || err.name === 'AuthSdkError') {
-        // remove expired tokens in storage
-        const removedTokens = [];
+        // remove expired token in storage
         const tokenStorage = storage.getStorage();
-        Object.keys(tokenStorage).forEach(key => {
-          const token = tokenStorage[key];
-          if (token && hasExpired(tokenMgmtRef, token)) {
-            delete tokenStorage[key];
-            removedTokens.push({ key, token });
-            clearExpireEventTimeout(tokenMgmtRef, key);
-          }
-        });
-        storage.setStorage(tokenStorage);
-        // emit removed events
-        if (!removedTokens.length) {
-          // tokens have been removed from other tabs
-          // still trigger removed event for downstream listeners
-          emitRemoved(tokenMgmtRef, ID_TOKEN_STORAGE_KEY);
-          emitRemoved(tokenMgmtRef, ACCESS_TOKEN_STORAGE_KEY);
+        const currentToken = tokenStorage[key];
+        if (currentToken && hasExpired(tokenMgmtRef, currentToken)) {
+          delete tokenStorage[key];
+          clearExpireEventTimeout(tokenMgmtRef, key);
+          storage.setStorage(tokenStorage);
+          emitRemoved(tokenMgmtRef, key, currentToken);
         } else {
-          removedTokens.forEach((key, token) => emitRemoved(tokenMgmtRef, key, token));
+          // token have been removed from other tabs
+          // still trigger removed event for downstream listeners
+          emitRemoved(tokenMgmtRef, key);
         }
-
+        
         err.tokenKey = key;
         emitError(tokenMgmtRef, err);
       }